Phase 4: Major Disconnect Test Consolidation
