--- tests/integration/test_disconnect_service_integration.py
@@ -1,366 +1,0 @@
-"""
-Integration tests for disconnect handling across multiple services.
-
-Tests the integration between services during disconnect scenarios using mocks
-to simulate various conditions and verify service interactions.
-"""
-
-import pytest
-from unittest.mock import Mock, patch, MagicMock, call
-from datetime import datetime
-import time
-
-from src.room_manager import RoomManager
-from src.game_manager import GameManager
-from src.services.auto_game_flow_service import AutoGameFlowService
-from src.services.broadcast_service import BroadcastService
-from src.services.session_service import SessionService
-from src.services.error_response_factory import ErrorResponseFactory
-from src.services.room_state_presenter import RoomStatePresenter
-from tests.helpers.socket_mocks import create_mock_socketio
-
-
-class TestDisconnectServiceIntegration:
-    """Test service integration during disconnect scenarios."""
-
-    def setup_method(self):
-        """Set up test fixtures with real services."""
-        self.room_manager = RoomManager()
-        self.game_manager = GameManager(self.room_manager)
-        self.session_service = SessionService()
-        self.error_factory = ErrorResponseFactory()
-        self.room_state_presenter = RoomStatePresenter(self.game_manager)
-
-        # Mock socketio for broadcast service using shared pattern
-        self.mock_socketio = create_mock_socketio()
-        self.broadcast_service = BroadcastService(
-            self.mock_socketio, self.room_manager, self.game_manager,
-            self.error_factory, self.room_state_presenter
-        )
-
-        self.auto_flow_service = AutoGameFlowService(
-            self.broadcast_service, self.game_manager, self.room_manager
-        )
-
-    def test_full_disconnect_workflow_integration(self):
-        """Test complete disconnect workflow across all services."""
-        room_id = "integration-room"
-        socket_id = "socket123"
-
-        # 1. Set up initial state
-        player_data = self.room_manager.add_player_to_room(room_id, "Alice", socket_id)
-        player_id = player_data['player_id']
-
-        session_data = {
-            'room_id': room_id,
-            'player_id': player_id,
-            'player_name': 'Alice'
-        }
-        self.session_service.create_session(socket_id, **session_data)
-
-        # 2. Simulate disconnect workflow
-
-        # Step 1: Mark player as disconnected in room
-        disconnect_success = self.room_manager.disconnect_player_from_room(room_id, player_id)
-        assert disconnect_success == True
-
-        # Step 2: Handle game impact
-        self.auto_flow_service.handle_player_disconnect_game_impact(room_id, player_id)
-
-        # Step 3: Broadcast updates
-        self.broadcast_service.broadcast_player_list_update(room_id)
-        self.broadcast_service.broadcast_room_state_update(room_id)
-
-        # Step 4: Clean up session
-        self.session_service.remove_session(socket_id)
-
-        # 3. Verify final state
-
-        # Player should be disconnected but preserved
-        connected_players = self.room_manager.get_connected_players(room_id)
-        assert len(connected_players) == 0
-
-        room_state = self.room_manager.get_room_state(room_id)
-        assert len(room_state['players']) == 1
-        player = list(room_state['players'].values())[0]
-        assert player['connected'] == False
-        assert player['name'] == 'Alice'
-
-        # Session should be removed
-        assert self.session_service.has_session(socket_id) == False
-
-        # Broadcasts should have been made
-        assert self.mock_socketio.emit.call_count >= 2  # At least player list and room state
-
-    def test_disconnect_with_broadcast_service_mocking(self):
-        """Test disconnect handling with mocked broadcast service to verify calls."""
-        room_id = "mock-broadcast-room"
-
-        # Set up players
-        alice_data = self.room_manager.add_player_to_room(room_id, "Alice", "socket1")
-        bob_data = self.room_manager.add_player_to_room(room_id, "Bob", "socket2")
-
-        # Mock broadcast service
-        mock_broadcast = Mock()
-        auto_flow_with_mock = AutoGameFlowService(
-            mock_broadcast, self.game_manager, self.room_manager
-        )
-
-        # Disconnect Alice
-        self.room_manager.disconnect_player_from_room(room_id, alice_data['player_id'])
-
-        # Trigger auto flow
-        auto_flow_with_mock.handle_player_disconnect_game_impact(room_id, alice_data['player_id'])
-
-        # Manual broadcast calls (simulating handler)
-        mock_broadcast.broadcast_player_list_update(room_id)
-        mock_broadcast.broadcast_room_state_update(room_id)
-
-        # Verify broadcast calls were made
-        mock_broadcast.broadcast_player_list_update.assert_called_once_with(room_id)
-        mock_broadcast.broadcast_room_state_update.assert_called_once_with(room_id)
-
-
-    def test_concurrent_disconnects_service_coordination(self):
-        """Test multiple simultaneous disconnects are handled correctly by services."""
-        room_id = "concurrent-room"
-
-        # Set up game with 4 players
-        players = []
-        for i, name in enumerate(['Alice', 'Bob', 'Charlie', 'Diana']):
-            player_data = self.room_manager.add_player_to_room(room_id, name, f"socket{i}")
-            players.append(player_data)
-
-        # Verify initial state
-        connected_players = self.room_manager.get_connected_players(room_id)
-        assert len(connected_players) == 4
-
-        # Disconnect multiple players
-        self.room_manager.disconnect_player_from_room(room_id, players[0]['player_id'])  # Alice
-        self.room_manager.disconnect_player_from_room(room_id, players[2]['player_id'])  # Charlie
-
-        # Verify intermediate state
-        connected_players = self.room_manager.get_connected_players(room_id)
-        assert len(connected_players) == 2
-        remaining_names = [p['name'] for p in connected_players]
-        assert 'Bob' in remaining_names
-        assert 'Diana' in remaining_names
-
-        # Trigger auto flow for each disconnect
-        self.auto_flow_service.handle_player_disconnect_game_impact(room_id, players[0]['player_id'])
-        self.auto_flow_service.handle_player_disconnect_game_impact(room_id, players[2]['player_id'])
-
-        # Verify final state consistency
-        room_state = self.room_manager.get_room_state(room_id)
-        assert len(room_state['players']) == 4  # All preserved
-
-        connected_players = self.room_manager.get_connected_players(room_id)
-        assert len(connected_players) == 2  # Only 2 connected
-
-
-
-class TestDisconnectBroadcastIntegration:
-    """Test broadcast service integration during disconnects."""
-
-    def setup_method(self):
-        """Set up test fixtures."""
-        self.room_manager = RoomManager()
-        self.game_manager = GameManager(self.room_manager)
-        self.error_factory = ErrorResponseFactory()
-        self.room_state_presenter = RoomStatePresenter(self.game_manager)
-        self.mock_socketio = create_mock_socketio()
-
-        self.broadcast_service = BroadcastService(
-            self.mock_socketio, self.room_manager, self.game_manager,
-            self.error_factory, self.room_state_presenter
-        )
-
-    def test_player_list_broadcast_after_disconnect(self):
-        """Test that player list broadcasts correctly reflect disconnected players."""
-        room_id = "broadcast-room"
-
-        # Add players
-        alice_data = self.room_manager.add_player_to_room(room_id, "Alice", "socket1")
-        bob_data = self.room_manager.add_player_to_room(room_id, "Bob", "socket2")
-
-        # Mock room state presenter to return expected format
-        expected_payload = {
-            'players': [
-                {'player_id': alice_data['player_id'], 'name': 'Alice', 'score': 0, 'connected': False},
-                {'player_id': bob_data['player_id'], 'name': 'Bob', 'score': 0, 'connected': True}
-            ],
-            'connected_count': 1,
-            'total_count': 2
-        }
-
-        # Mock the room state and presenter
-        with patch.object(self.room_manager, 'get_connected_players') as mock_connected:
-            mock_connected.return_value = [{'player_id': bob_data['player_id'], 'name': 'Bob', 'score': 0, 'connected': True}]
-
-            with patch.object(self.room_state_presenter, 'create_player_list_update') as mock_presenter:
-                mock_presenter.return_value = expected_payload
-
-                # Broadcast player list update
-                self.broadcast_service.broadcast_player_list_update(room_id)
-
-                # Verify correct data was used
-                mock_presenter.assert_called_once()
-                mock_connected.assert_called_once_with(room_id)
-
-                # Verify broadcast was made with correct event
-                self.mock_socketio.emit.assert_called_once_with(
-                    'player_list_updated', expected_payload, room=room_id
-                )
-
-    def test_room_state_broadcast_excludes_sensitive_data(self):
-        """Test that room state broadcasts properly filter sensitive data."""
-        room_id = "sensitive-room"
-
-        # Add player and set up game
-        self.room_manager.add_player_to_room(room_id, "Alice", "socket1")
-
-        # Mock room state with sensitive data
-        mock_room_state = {
-            'players': {'player1': {'name': 'Alice', 'connected': False}},
-            'game_state': {
-                'phase': 'responding',
-                'current_prompt': {
-                    'id': 'prompt1',
-                    'prompt': 'Test prompt',
-                    'model': 'gpt-4',
-                    'ai_response': 'SECRET_RESPONSE'  # Should be filtered
-                }
-            }
-        }
-
-        # Mock presenter to return safe data
-        safe_game_state = {
-            'phase': 'responding',
-            'current_prompt': {
-                'id': 'prompt1',
-                'prompt': 'Test prompt',
-                'model': 'gpt-4'
-                # ai_response should be filtered out
-            }
-        }
-
-        with patch.object(self.room_manager, 'get_room_state') as mock_room_state_getter:
-            mock_room_state_getter.return_value = mock_room_state
-
-            with patch.object(self.room_state_presenter, 'create_safe_game_state') as mock_presenter:
-                mock_presenter.return_value = safe_game_state
-
-                # Broadcast room state
-                self.broadcast_service.broadcast_room_state_update(room_id)
-
-                # Verify presenter was called to filter sensitive data
-                mock_presenter.assert_called_once_with(mock_room_state, room_id)
-
-                # Verify broadcast used filtered data
-                self.mock_socketio.emit.assert_called_once_with(
-                    'room_state_updated', safe_game_state, room=room_id
-                )
-
-
-
-class TestDisconnectMemoryManagement:
-    """Test memory management during disconnect scenarios."""
-
-    def setup_method(self):
-        """Set up test fixtures."""
-        self.room_manager = RoomManager()
-        self.session_service = SessionService()
-
-    def test_session_memory_cleanup(self):
-        """Test that sessions are properly cleaned up and don't leak memory."""
-        initial_session_count = len(self.session_service.get_all_sessions())
-
-        # Create multiple sessions
-        socket_ids = [f"socket_{i}" for i in range(10)]
-        for socket_id in socket_ids:
-            self.session_service.create_session(
-                socket_id,
-                room_id=f"room_{socket_id}",
-                player_id=f"player_{socket_id}",
-                player_name=f"Player{socket_id}"
-            )
-
-        # Verify sessions created
-        assert len(self.session_service.get_all_sessions()) == initial_session_count + 10
-
-        # Remove all sessions (simulating disconnects)
-        for socket_id in socket_ids:
-            self.session_service.remove_session(socket_id)
-
-        # Verify all sessions removed
-        final_session_count = len(self.session_service.get_all_sessions())
-        assert final_session_count == initial_session_count
-
-    def test_room_state_preservation_during_disconnects(self):
-        """Test that room state is preserved correctly during multiple disconnects."""
-        room_id = "preservation-room"
-
-        # Add multiple players
-        players = []
-        for i in range(5):
-            player_data = self.room_manager.add_player_to_room(room_id, f"Player{i}", f"socket{i}")
-            players.append(player_data)
-
-        initial_room_state = self.room_manager.get_room_state(room_id)
-        initial_player_count = len(initial_room_state['players'])
-
-        # Disconnect all but one player
-        for i in range(4):
-            self.room_manager.disconnect_player_from_room(room_id, players[i]['player_id'])
-
-        # Verify room state preservation
-        final_room_state = self.room_manager.get_room_state(room_id)
-
-        # All players should still be in room state
-        assert len(final_room_state['players']) == initial_player_count
-
-        # Only one should be connected
-        connected_players = self.room_manager.get_connected_players(room_id)
-        assert len(connected_players) == 1
-        assert connected_players[0]['name'] == 'Player4'
-
-        # Disconnected players should have preserved data
-        all_players = list(final_room_state['players'].values())
-        disconnected_players = [p for p in all_players if not p['connected']]
-        assert len(disconnected_players) == 4
-
-        # Verify data integrity
-        for i, player in enumerate(disconnected_players):
-            assert player['name'] in [f"Player{j}" for j in range(4)]
-            assert player['score'] == 0  # Score preserved
-
-    def test_no_memory_leaks_in_repeated_disconnect_reconnect(self):
-        """Test no memory leaks occur during repeated disconnect/reconnect cycles."""
-        room_id = "cycle-room"
-
-        # Simulate multiple disconnect/reconnect cycles
-        for cycle in range(10):
-            # Add player
-            player_data = self.room_manager.add_player_to_room(room_id, f"Player{cycle}", f"socket{cycle}")
-
-            # Verify player added
-            connected = self.room_manager.get_connected_players(room_id)
-            assert len(connected) == 1
-
-            # Disconnect player
-            self.room_manager.disconnect_player_from_room(room_id, player_data['player_id'])
-
-            # Verify disconnected but preserved
-            room_state = self.room_manager.get_room_state(room_id)
-            assert len(room_state['players']) == 1
-
-            connected = self.room_manager.get_connected_players(room_id)
-            assert len(connected) == 0
-
-            # Remove player completely (simulate cleanup)
-            self.room_manager.remove_player_from_room(room_id, player_data['player_id'])
-
-        # Verify no accumulated state
-        final_room_state = self.room_manager.get_room_state(room_id)
-        if final_room_state:  # Room might be auto-cleaned
-            assert len(final_room_state['players']) == 0
