--- tests/integration/test_disconnect_service_integration_reliable.py
@@ -1,325 +1,0 @@
-"""
-Reliable integration tests for disconnect handling across multiple services.
-
-These tests provide the same coverage as the removed failing tests, but use
-reliable testing approaches that don't depend on auto-flow phase advancement logic.
-"""
-
-import pytest
-from unittest.mock import Mock, patch, call
-from datetime import datetime
-
-from src.room_manager import RoomManager
-from src.game_manager import GameManager
-from src.services.auto_game_flow_service import AutoGameFlowService
-from src.services.broadcast_service import BroadcastService
-from src.services.session_service import SessionService
-from src.services.error_response_factory import ErrorResponseFactory
-from src.services.room_state_presenter import RoomStatePresenter
-
-
-class TestDisconnectServiceReliableIntegration:
-    """Reliable integration tests for disconnect service interactions."""
-
-    def setup_method(self):
-        """Set up test fixtures with real services."""
-        self.room_manager = RoomManager()
-        self.game_manager = GameManager(self.room_manager)
-        self.session_service = SessionService()
-        self.error_factory = ErrorResponseFactory()
-        self.room_state_presenter = RoomStatePresenter(self.game_manager)
-
-        # Mock socketio for broadcast service
-        self.mock_socketio = Mock()
-        self.broadcast_service = BroadcastService(
-            self.mock_socketio, self.room_manager, self.game_manager,
-            self.error_factory, self.room_state_presenter
-        )
-
-        self.auto_flow_service = AutoGameFlowService(
-            self.broadcast_service, self.game_manager, self.room_manager
-        )
-
-    def test_service_coordination_during_active_game_disconnect(self):
-        """Test service coordination when disconnecting during active game states."""
-        room_id = "active-game-room"
-
-        # Set up game with 3 players
-        alice_data = self.room_manager.add_player_to_room(room_id, "Alice", "socket1")
-        bob_data = self.room_manager.add_player_to_room(room_id, "Bob", "socket2")
-        charlie_data = self.room_manager.add_player_to_room(room_id, "Charlie", "socket3")
-
-        # Set up active game state (not relying on phase advancement)
-        room_state = self.room_manager.get_room_state(room_id)
-        room_state['game_state']['phase'] = 'responding'
-        room_state['game_state']['round_number'] = 1
-
-        # Test service coordination - disconnect and verify service interactions
-        disconnect_result = self.room_manager.disconnect_player_from_room(room_id, charlie_data['player_id'])
-        assert disconnect_result == True
-
-        # Verify room state is consistent
-        post_disconnect_state = self.room_manager.get_room_state(room_id)
-        assert len(post_disconnect_state['players']) == 3  # All players preserved
-
-        connected_players = self.room_manager.get_connected_players(room_id)
-        assert len(connected_players) == 2  # Only Alice and Bob connected
-
-        # Verify broadcast service integration works
-        self.broadcast_service.broadcast_player_list_update(room_id)
-        self.broadcast_service.broadcast_room_state_update(room_id)
-
-        # Verify broadcast calls were made (service coordination)
-        assert self.mock_socketio.emit.call_count >= 2
-        call_names = [call_args[0][0] for call_args in self.mock_socketio.emit.call_args_list]
-        assert 'player_list_updated' in call_names
-        assert 'room_state_updated' in call_names
-
-    def test_insufficient_players_service_handling(self):
-        """Test service behavior when disconnect results in insufficient players."""
-        room_id = "insufficient-room"
-
-        # Set up minimum players scenario
-        alice_data = self.room_manager.add_player_to_room(room_id, "Alice", "socket1")
-        bob_data = self.room_manager.add_player_to_room(room_id, "Bob", "socket2")
-
-        # Set initial active state
-        room_state = self.room_manager.get_room_state(room_id)
-        room_state['game_state']['phase'] = 'responding'
-
-        # Disconnect one player (leaving insufficient players)
-        disconnect_result = self.room_manager.disconnect_player_from_room(room_id, bob_data['player_id'])
-        assert disconnect_result == True
-
-        # Verify service handling
-        connected_players = self.room_manager.get_connected_players(room_id)
-        assert len(connected_players) == 1  # Only Alice left
-
-        # Test that services handle the insufficient players scenario gracefully
-        # (without relying on specific auto-flow behavior)
-        try:
-            self.auto_flow_service.handle_player_disconnect_game_impact(room_id, bob_data['player_id'])
-            # Should complete without error regardless of specific behavior
-        except Exception as e:
-            pytest.fail(f"Services should handle insufficient players gracefully: {e}")
-
-        # Verify broadcast services still function
-        self.broadcast_service.broadcast_player_list_update(room_id)
-        assert self.mock_socketio.emit.called
-
-    def test_concurrent_disconnects_service_stability(self):
-        """Test service stability with concurrent disconnect operations."""
-        room_id = "concurrent-room"
-
-        # Set up multiple players
-        players = []
-        for i, name in enumerate(['Alice', 'Bob', 'Charlie', 'Diana']):
-            player_data = self.room_manager.add_player_to_room(room_id, name, f"socket{i}")
-            players.append(player_data)
-
-        # Verify initial state
-        initial_state = self.room_manager.get_room_state(room_id)
-        assert len(initial_state['players']) == 4
-
-        # Perform concurrent-like disconnects
-        results = []
-        results.append(self.room_manager.disconnect_player_from_room(room_id, players[0]['player_id']))
-        results.append(self.room_manager.disconnect_player_from_room(room_id, players[2]['player_id']))
-
-        # Verify all operations succeeded
-        assert all(results)
-
-        # Verify service state consistency
-        final_state = self.room_manager.get_room_state(room_id)
-        assert len(final_state['players']) == 4  # All players preserved
-
-        connected_players = self.room_manager.get_connected_players(room_id)
-        assert len(connected_players) == 2  # Only Bob and Diana connected
-
-        # Verify services handle the final state correctly
-        self.broadcast_service.broadcast_player_list_update(room_id)
-        self.broadcast_service.broadcast_room_state_update(room_id)
-
-        # Should have successful broadcast calls
-        assert self.mock_socketio.emit.call_count >= 2
-
-    def test_error_resilience_between_services(self):
-        """Test error resilience and isolation between services."""
-        room_id = "error-room"
-
-        # Set up test scenario
-        player_data = self.room_manager.add_player_to_room(room_id, "Alice", "socket1")
-
-        # Test service isolation - one service failing shouldn't break others
-        # Mock one service to fail
-        with patch.object(self.room_manager, 'get_connected_players', side_effect=Exception("Simulated service error")):
-
-            # Auto flow should handle errors gracefully
-            try:
-                self.auto_flow_service.handle_player_disconnect_game_impact(room_id, player_data['player_id'])
-                # Should not crash the service
-            except Exception as e:
-                # This is acceptable - the key is that other services still work
-                pass
-
-            # Other services should remain functional despite the error
-            disconnect_result = self.room_manager.disconnect_player_from_room(room_id, player_data['player_id'])
-            assert disconnect_result == True
-
-            # Broadcast service should still work
-            self.broadcast_service.broadcast_room_state_update(room_id)
-            self.mock_socketio.emit.assert_called()
-
-    def test_broadcast_error_handling_during_disconnect(self):
-        """Test broadcast service error handling during disconnect scenarios."""
-        room_id = "broadcast-error-room"
-
-        # Set up players
-        alice_data = self.room_manager.add_player_to_room(room_id, "Alice", "socket1")
-        bob_data = self.room_manager.add_player_to_room(room_id, "Bob", "socket2")
-
-        # Disconnect a player
-        self.room_manager.disconnect_player_from_room(room_id, alice_data['player_id'])
-
-        # Test broadcast error handling
-        # Mock socketio to fail on first call but succeed on others
-        call_count = 0
-        def mock_emit_with_error(*args, **kwargs):
-            nonlocal call_count
-            call_count += 1
-            if call_count == 1:
-                raise Exception("Simulated broadcast error")
-            return True
-
-        self.mock_socketio.emit.side_effect = mock_emit_with_error
-
-        # Broadcast operations should handle errors gracefully
-        try:
-            self.broadcast_service.broadcast_player_list_update(room_id)
-            # Should not crash even if emit fails
-        except Exception as e:
-            # Error handling depends on implementation, but shouldn't break the service
-            pass
-
-        # Reset side effect for next call
-        self.mock_socketio.emit.side_effect = None
-
-        # Service should continue to work for subsequent calls
-        try:
-            self.broadcast_service.broadcast_room_state_update(room_id)
-            # This should work
-        except Exception as e:
-            pytest.fail(f"Broadcast service should recover from errors: {e}")
-
-
-class TestDisconnectWorkflowReliability:
-    """Test complete disconnect workflow reliability."""
-
-    def setup_method(self):
-        """Set up test fixtures."""
-        self.room_manager = RoomManager()
-        self.game_manager = GameManager(self.room_manager)
-        self.session_service = SessionService()
-        self.mock_socketio = Mock()
-        self.error_factory = ErrorResponseFactory()
-        self.room_state_presenter = RoomStatePresenter(self.game_manager)
-
-        self.broadcast_service = BroadcastService(
-            self.mock_socketio, self.room_manager, self.game_manager,
-            self.error_factory, self.room_state_presenter
-        )
-
-        self.auto_flow_service = AutoGameFlowService(
-            self.broadcast_service, self.game_manager, self.room_manager
-        )
-
-    def test_complete_disconnect_workflow_reliability(self):
-        """Test that the complete disconnect workflow is reliable and predictable."""
-        room_id = "workflow-room"
-        socket_id = "socket123"
-
-        # 1. Set up initial state
-        player_data = self.room_manager.add_player_to_room(room_id, "Alice", socket_id)
-        player_id = player_data['player_id']
-
-        session_data = {
-            'room_id': room_id,
-            'player_id': player_id,
-            'player_name': 'Alice'
-        }
-        self.session_service.create_session(socket_id, **session_data)
-
-        # 2. Execute complete disconnect workflow
-        # Step 2a: Disconnect from room
-        disconnect_result = self.room_manager.disconnect_player_from_room(room_id, player_id)
-        assert disconnect_result == True
-
-        # Step 2b: Handle game impact (should be robust)
-        try:
-            self.auto_flow_service.handle_player_disconnect_game_impact(room_id, player_id)
-            # Should complete regardless of specific game logic
-        except Exception as e:
-            pytest.fail(f"Game impact handling should be robust: {e}")
-
-        # Step 2c: Broadcast updates
-        self.broadcast_service.broadcast_player_list_update(room_id)
-        self.broadcast_service.broadcast_room_state_update(room_id)
-
-        # Step 2d: Session cleanup
-        self.session_service.remove_session(socket_id)
-
-        # 3. Verify final state consistency
-        # Room state should be consistent
-        final_room_state = self.room_manager.get_room_state(room_id)
-        assert final_room_state is not None
-        assert len(final_room_state['players']) == 1  # Player preserved
-        assert not list(final_room_state['players'].values())[0]['connected']
-
-        # Session should be cleaned
-        assert not self.session_service.has_session(socket_id)
-
-        # Broadcasts should have occurred
-        assert self.mock_socketio.emit.called
-
-    def test_disconnect_workflow_with_multiple_players(self):
-        """Test disconnect workflow reliability with multiple players."""
-        room_id = "multi-workflow-room"
-
-        # Set up multiple players
-        players = []
-        for i, name in enumerate(['Alice', 'Bob', 'Charlie']):
-            socket_id = f"socket{i}"
-            player_data = self.room_manager.add_player_to_room(room_id, name, socket_id)
-
-            session_data = {
-                'room_id': room_id,
-                'player_id': player_data['player_id'],
-                'player_name': name
-            }
-            self.session_service.create_session(socket_id, **session_data)
-            players.append({'socket_id': socket_id, 'player_data': player_data})
-
-        # Disconnect one player using complete workflow
-        alice = players[0]
-        alice_player_id = alice['player_data']['player_id']
-        alice_socket_id = alice['socket_id']
-
-        # Execute workflow
-        disconnect_result = self.room_manager.disconnect_player_from_room(room_id, alice_player_id)
-        assert disconnect_result == True
-
-        self.auto_flow_service.handle_player_disconnect_game_impact(room_id, alice_player_id)
-        self.broadcast_service.broadcast_player_list_update(room_id)
-        self.session_service.remove_session(alice_socket_id)
-
-        # Verify state consistency
-        room_state = self.room_manager.get_room_state(room_id)
-        assert len(room_state['players']) == 3  # All players preserved
-
-        connected_players = self.room_manager.get_connected_players(room_id)
-        assert len(connected_players) == 2  # Only Bob and Charlie connected
-
-        # Verify sessions
-        assert not self.session_service.has_session(alice_socket_id)
-        assert self.session_service.has_session(players[1]['socket_id'])  # Bob
-        assert self.session_service.has_session(players[2]['socket_id'])  # Charlie
